parameter-framework: Added string parameters
BZ: 12819

String parameters are specified in the XML structure with a MaxLength attribute.
MaxLength corresponds to the maximum amount of characters the string parameter
can handle.
String parameters are expressend in the form of tokens, that is they don't
contain any sort of space characters.
For now, string parameter arrays are not supported.

Change-Id: Ifbb006e3d2e3077535c32c0fd579cc04872a82b0
Signed-off-by: Patrick Benavoli <patrickx.benavoli@intel.com>
Reviewed-on: http://android.intel.com:8080/22319
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/26778
Reviewed-by: Barthes, FabienX <fabienx.barthes@intel.com>