PFW: Fixed bug on fixed point parameter types
BZ: 29464

When reading from the blackboard a value of type fixed point
parameter (size < 32 bits), the sign extend was not taken into account:
- Added a call to signExtend() before the conversion,
- Updated the display as well (precision support),
- Set up the notation type (fixed) to prevent the scientific notation display.

Compilation on target would fail because of the use of an ambigous
function : log10. Changed the notation of the number to double for the
parameter of this function to force binding to: double log10(double).

Change-Id: Ibb77259b84db33e00971ba662600547522cf9bb1
Signed-off-by: Frederic Boisnard <fredericx.boisnard@intel.com>
Reviewed-on: http://android.intel.com:8080/41649
Reviewed-by: De Chivre, RenaudX <renaudx.de.chivre@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: Benavoli, PatrickX <patrickx.benavoli@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>