PFW: Fix precision error on FixedPointParameters
BZ: 36789

When exporting the settings to an XML file and importing this file right after,
an error occurs on some fixed point parameters.  The values of these
parameters are incorrectly saved into the output file.

This bug is due to a precision error when converting the raw value from the
blackboard. The setprecision() manipulator rounds the result instead of
truncating it, sometimes leading to an overflow.  The fix consists in
performing a truncation after the last significant digit.

Change-Id: I70c59bffabfd61f1ce95a8fb4cef33273a326eaf
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>
Reviewed-on: http://android.intel.com:8080/50935
Reviewed-by: Benavoli, Patrick <patrick.benavoli@intel.com>
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Reviewed-by: Denneulin, Guillaume <guillaume.denneulin@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>