PFW: Check if value provided is exactly an integer
BZ: 50802

Previously, if a fixed point value was provided for an integer parameter, there was no error.
This was due to the results of strtol/strtoll which was not correctly checked (it would only fail
if no digit was provided, but would succeed even if letters were appended after/before the digits).
So for fixed point values, only the first part of the string was fetched, giving the impression of a truncation at the '.' character .
Now the whole string shall be completly parsed for the conversion to succeed.

Change-Id: Ia907e8cc05e5639e9a8bbd7204f46df5bdfa394a
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>
Reviewed-on: http://android.intel.com:8080/60743
Reviewed-by: Rocard, KevinX <kevinx.rocard@intel.com>
Reviewed-by: Benavoli, Patrick <patrick.benavoli@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Mendi, EduardoX <eduardox.mendi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>