PFW: Show pending configurations in status command
BZ: 56118

Update the "status" command of the PFW to display the current pending
configurations for all domains, that is to say configurations that
are waiting the next call to "applyConfiguration()" to be applied.
This improvement will help developers to better understand
the current state of the PFW, in particular when dealing with HAL
modifications.

Change-Id: I6300620ad8cedcb534fb2859b57b7ab2988fa618
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>
Reviewed-on: http://android.intel.com:8080/64796
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: Benavoli, Patrick <patrick.benavoli@intel.com>
Reviewed-by: Rocard, KevinX <kevinx.rocard@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>