PFW: Correct truncated criterion log
BZ: 85746

Create a log fonction to log a string table. Use it to call the log
function for each criterion.

Context:
When the PFW starts, it logs the criteria, their states and possible
values. All criterion logs where concatenate in a string and logged.

But if there is a lot of criteria, the 512 chars long log buffer might
be overfload and the log is truncated.

Change-Id: I72c89b25de9314b714763b3ac71a7d0cb9698910
Signed-off-by: Kevin Rocard <kevinx.rocard@intel.com>
Reviewed-on: http://android.intel.com:8080/90089
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>