[PFW] Fix display error when using getConfigurationParameter with BooleanParameters
BZ: 96140

When using getConfigurationParameter on a BooleanParameter, the displayed value is
duplicated on the output if the configuration is applicable.

The method CParameterMgr::accessConfigurationValue() accesses the element twice
when the configuration is the last one applied (in order to update the parameter value
in both the configuration blackboard and the main blackboard).
The output result is managed using a string passed as reference
to ParameterMgr::accessValue() and then to CBooleanParameterType::fromBlackboard().
In the latest method, the boolean value is appended to the output string, however no
check is performed to make sure that the string is empty before being used.
As the method is called twice with the same output string, we end up with a dupplicated
output.
This fix updates CBooleanParameterType::fromBlackboard() in order to make sure that
the output string is overwritten with the current result.

Change-Id: I394366910b1ddc0c8bbe570aa357e007745613db
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>
Reviewed-on: http://android.intel.com:8080/98648
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Denneulin, Guillaume <guillaume.denneulin@intel.com>
Reviewed-by: Benavoli, Patrick <patrick.benavoli@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>