PFW: Remove useless export date on top of settings XML file
BZ: 105122

On line 3 of configurable domains xml files, a comment was added
to specify date of files creation. This date makes all trivial
merges on this file to fail, making rebases a harder task.

Remove this comment as this kind of files are changed very often
and as all modifications to this file are tracked with git.

Change-Id: I4e8aedf95bbdea03137709f91242179d6acab5f3
Signed-off-by: Guillaume Denneulin <guillaume.denneulin@intel.com>
Reviewed-on: http://android.intel.com:8080/105823
Reviewed-by: Rocard, KevinX <kevinx.rocard@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>