[Coverage] Raise duplicated criterion specialised exception
BZ: 115218

When an element is requested to adopt multiple times the same element, the
"DuplicatedChildError" exception is raised.
However, in case of PFW restart, existing criteria are recreated, which
is throwing the generic "DuplicatedChildError" exception.

As such event need to be detected, specialise the "DuplicatedChildError"
exception in "DuplicatedCriterionError" in order to catch the event
easily.

Change-Id: Ib0cca1d318233ca0b2f69fd277740ce8f5a550ad
Signed-off-by: Kevin Rocard <kevinx.rocard@intel.com>
Reviewed-on: http://android.intel.com:8080/112712
Reviewed-by: Denneulin, Guillaume <guillaume.denneulin@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>