[Makefile] Append "_host" to host modules names
BZ: 115743

In some parameter framework makefiles, host and target modules (with the
same name) are defined in the same file.
External prebuilt framework properly handle this case when modules are
defined in different Android.mk, but there is a limitation when modules
are defined in same Android.mk.

Append _host to all host modules.

Change-Id: Ib594dabd86617cc40993066338f095b49afb2ea2
Signed-off-by: Kevin Rocard <kevinx.rocard@intel.com>
Reviewed-on: http://android.intel.com:8080/113537
Reviewed-by: Denneulin, Guillaume <guillaume.denneulin@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>