Fix issue in PFW when loading host libraries
BZ: 127874

Default entry point for plugins libraries must follow the pattern
"lib*-subsystem.so". However, when compiling on host, the plugins
are named "lib*-subsystem_host.so", so the PFW is unable to load
them.

This patch modifies the way plugins symbols are detected, so that
host libraries can be loaded successfully.

Change-Id: I079162c3e47a975641b3d7cb1fc054f95c86028e
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>
Reviewed-on: http://android.intel.com:8080/122812
Reviewed-by: Rocard, KevinX <kevinx.rocard@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>