Adds default value pair for inclusive criterion
BZ: 99956

Inclusive criterion type does not allow to check if the bitfield
is empty, i.e., if zero is not available.

This patch adds the default value pair "none, 0" to allow
check either if the criterion "Is" 0, and establishes consistency
with inclusion rule. It also allows to convert a concatenated
chain of string value, where "|" is considered as AND rule into
the corresponding numerical value.

Change-Id: I4ee7c0f69228a727964ab80050ac12c6babd4f58
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Signed-off-by: Jorge Quintero <jorge.quintero@intel.com>
Reviewed-on: http://android.intel.com:8080/138518
Reviewed-by: Centelles, Sylvain <sylvain.centelles@intel.com>
Tested-by: Barthes, FabienX <fabienx.barthes@intel.com>