Allow concurent generation of XML routing files
BZ: 145938

XML generation is impossible when multiple instances of the script
hostDomainGeneration are run (this is typically the case with the
current Android Make process). This is because the way the ports
used to communicate with test-platform and parameter-framework are
chosen is not protected agains concurrent executions.

This patch protects the initialization of the test-platform and
parameter-framework agains concurrent executions using a flock.
In particular, it ensures that the ports chosen are reserved to
a specific process.

Change-Id: I3d8fee2b20313814c79f5390851ad57dc696cc28
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>