Fix hostDomainGenerator socket-checking mechanism for Ubuntu 13.04
BZ: 148729

The script hostDomainGenerator.sh doesn't work fine on Ubuntu 13.04.
The processing is based on the external tool "ss" to monitor opened
sockets, in order to check if existing instances of test-platform
are currently started. It appears that the behavior of this tool
is different in the latest Ubuntu release: requests of opened sockets
always return multiple false results. In consequence, the script
cannot find any available port to start a new test-platform instance.

This patch updates the way opened sockets are checked, using "grep"
instead of the "ss" filter mechanism.

Change-Id: I2e1245c524d021d99389a0ec028105ed1f706666
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>