hostDomainGenerator.sh: fix lock permissions
BZ: 149851

The hostDomainGenerator.sh script uses a lockfile to protect some
operations from concurrent execution. This mechanism uses a file
that is created if it does not already exist. In this case, the file
permissions are too stricts, and other users are unable to run the script
anymore.

This patch changes the permission on the lockfile, allowing other
users to access it.

Change-Id: Iec250a4022cb8f97973fa92020b27dc0e83890c6
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>