remote-process: allow reading commands from stdin
BZ: 154316

The remote-process utility was only able to send one command at a time.

This patch adds a new usage: if the command is omitted, remote-process
reads stdin and sends each line as commands.  It stops whenever it
encounters an error of any kind (internal or returned by the server).

Example of the new usage:
remote-process localhost 5000 < commands.txt

Change-Id: I2478668807b5541544c5752e68e112c23c85af15
Signed-off-by: David Wagner <david.wagner@intel.com>