remote-processor: configure the socket with the TCP_NODELAY option
BZ: 154316

The "man 7 tcp" page speaks for itself:
TCP_NODELAY
  If  set, disable the Nagle algorithm.  This means that segments are always
sent as soon as possible, even if there is only a small amount of data.  When
not set, data is buffered until there is a  sufficient amount to send out,
thereby avoiding the frequent sending of small packets, which results in poor
utilization of the network.

The behaviour described above if this option is not set introduces a lot of
overhead when several commands are sent with remote-processor because they
are not sent in burst mode: instead, remote-processor wait for the answer to
each command.

See "man 7 tcp" for all the details.

Change-Id: I6c8ede4428629a16e9a2927d73a2897873a57d37
Signed-off-by: David Wagner <david.wagner@intel.com>