Add integer conversion that can take care of sign extension
BZ: 163707

When using signed integers of size 8 or 16, the bit of sign needs to
be propagated if the API called by the plugin wait for "int" type

Add the virtual toInteger function in CTypeElement for each parameter
type to take care of conversion to "int". Specialize this function in
the class CIntegerParameterType to take care of sign extension.

Change-Id: I41183dccbcc21212299d1dde86b3ad4ba8432ce4
Signed-off-by: Guillaume Denneulin <guillaume.denneulin@intel.com>