Allow starting test-platform as a Daemon
BZ: 151780

Currently, test-platform does not return when it is launched. In
particular, one has to try to connect to it in order to know if it
has correctly been initialized.

Added a new '-d' option to test-platform to start it as a daemon:

    test-platform [-d] <file path> [port number, default 5001]

When test-platform is started as a daemon, a child process is created
and the main process immediatly returns its status.

Change-Id: I70a33691909c958904cf50d156a563b998f92657
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>