Make hostDomainGenerator.sh verbose by default
BZ: 151780

The option --verbose must be provided in order to make hostDomainGenerator.sh
verbose. However, this option is very usefull when the generation fails during
a build, and it would be nice to have it enabled by default.

Make this option the default one, and add a nonverbose option instead.

Change-Id: I59195b7a28e25a5f18871d6baf2ae4f1310f49d7
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>