Fix "Argument list too long" bug during XML generation
BZ: 177060

When the initial XML file for tuning is too big, the XML generation
fails with the message: "Argument list too long". This is because the
generation relies on the PFW command "setDomainsWithSettingsXML", which
takes XML data as string argument. There is a limitation concerning
the size of this string, leading to the current issue.

This patch replaces the call to "setDomainsWithSettingsXML" by
"importDomainsWithSettingsXML", which takes a file as argument instead
of a string. There is no limitation on the file size.

Change-Id: Ieb12e6ba1391545021cd3f1e867dee0d28425d19
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>