XmlValidator tool
This tool can be used to validate a set of xml files against w3c
schemas.

It scans a directory recursively and checks for .xml files in it.
For each file, it extracts the current schema file based on a convention
(the corresponding schema file must have the same name as the root node)
which is used in the parameter-framework.

Change-Id: I27e32028e39a450ed18f86cb095348f73d59ade5