Local .xsd files for config validation
BZ: 184054

In order to verify that xml files are valid, the parameter framework
check them with .xsd files.
In those files, there are references towards base schemas on the
internet. Which means that every validation downloads a copy of
the base file via http.

To avoid doing http requests, we can use the file locally.
This patchs changes from remote file to local.

Change-Id: Ide48dba0000db7c97e1d269ee5573ac8afe7e4d4
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>