Test-platform enable/disable xml validation
BZ: 184054

test-platform has no way to tell to the parameter
framework that it wants the parameter framework to
validate xmls on loading.

New command added to test-platform executable.
setValidateSchemasOnStart true|false
    If set to true, the parameter framework will try
    to validate the .xmls with the .xsds located next to them.
    If set to false, it will never attempt to validate.

SetValidateSchemas is false by default to allow backward compatibility.

Note(a): This assumes that the script is able to
find the .xsd files (which should be in /tmp/Schemas/).

Change-Id: I7a5b0bc09d31e6647d8c631380d31503666e7e7b
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>