Added copying of xsd files to host build output
BZ: 188323

In order to validate the .xml parameter framework files, we use XSDs.
Those must be in the same directory as the .xml files when they
are validated.

This patch copies the .xsd files towards a directory that
hostDomainGenerator knows, so that he can use the files and
copy them towards /tmp.

Change-Id: I938fc3791006777ff65c96d4259596111f93d7b8
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>