Fix xmlserializer makefile for schema validation
BZ: 188323

For schema validation features, we are using a special version of libxml2 called
libxml2-schemas (which contains the SCHEMAS feature).
This is just a different static library build with that feature only for
the parameter-framework.
The problem with that, is that the include/xmlversion.h does not
define the #LIBXML_SCHEMAS_ENABLED flag.
So when including it in our project, the flag is not defined.
This results in unexpected behaviour since validation was active in the
static library, but never called because the flag was not defined.

As a fix, we are defining the LIBXML_SCHEMAS_ENABLED flag in the makefile.

Change-Id: Id56ae5b668e587171b11e7607f2603ab5e3341a1
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>