Updated hostDomainGenerator.sh schema validation
BZ: 188323

In order to validate xml files, we need the corresponding schemas to be
next to the xml files. Those are put by the script in /tmp.

This patch add the copying of the schema files to /tmp.
I also adds an optional command line argument to disable/enable validation.
This is useful for platform specific enabling/disabling validation.
It can be used by calling hostDomainGenerator.sh --validate.

Change-Id: I7da65f230e5abf5b1d18c5cd6cf11b67dae5ed76
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>