FixedPointParameter: Removed precision notion
BZ: 197723

Significant digits were computed in order to prevent odd behaviour when printing
a converted fixed point parameter. After converting a number from fixed
point towards a double, we had to round up/down in order to print the
double to the user (or towards a file).
The printing showed only the significant digits, in other words, the
digits which can vary from 0 to 9. This caused unpredictable behaviour
when rounding using the setPrecision() function.

This patch removes the 'precision' notion related to fixed points in
order to have a predictable behaviour by displaying all fractional
digits.

Change-Id: Id5a9b5b8a80d5a96bcbbb6af9213770a05a70083
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>