Removed phony package to copy .xsd files
BZ: 192861

When hostDomainGenerator.sh is called by the build system, it needs to
copy schema files to /tmp/tmp.XXXXXXX. This is because
parameter-framework needs to have the schema files next to the .xml
files in order to apply validation.

In order to do that, hostDomainGenerator.sh needs to know where he can
copy the .xsd files from.

Currently, the .xsd files are copied from
$HOST_OUT/etc/parameter-framework/Schemas via a dedicated makefile.
This used a phony package to group all the files that have to be copied.

The problem is that phony packages the phony package is not calling the
prebuildt rules we need. So the files we want to export to
/etc/parameter-framework/Schemas are not copied.

This patch removes the phony package in order to produce the expected
behaviour.

Change-Id: I96da56f85a8c3d512fcfb97244181dc6c1cb9401
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>