Fix bug when using setConfigurationParameter command
BZ: 189941

There is an issue when trying to set parameter value in a
configuration using the setConfigurationParameter command.
The error returned is "Unable to forward synchronize
configurable element, Failed to write value in mixer control".

The root cause is a recent update concerning the
auto synchronization mode which is now turned on by default.
The setConfigurationParameter command first updates the local
Configuration blackboard, and then the Main blackboard if this
Configuration is the one applicable for the domain. However,
this local blackboard should not be synchronized, only the
main blackboard should be. This patch deactivates synchronization
for the local blackboard.

Change-Id: I7ca71806309d19ae482399f474d32a0c1b70ae6c
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>