Android.mk hack for compiling libxml with schemas enabled
The "Schema" feature is enabled in the official libxml2.
We use that feature to validate the xml files given to the
parameter-framework.
In the libxml2 aosp version, this feature is not enabled.

This patch adds a new target, libxml2-schemas, which contains the
"Schemas" feature.

This should be removed when the "Schema" feature is integrated into the
aosp's libxml2 version.

Change-Id: Ida72de36e7fd196063a332ee70b30c0f0cb69024
Signed-off-by: David Wagner <david.wagner@intel.com>
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>