Delete subsystem instances *before* plugin unload
BZ: 208504

Before this patch there was a crash in CElement destructor in Ubuntu upon
destuction of parameter framework instance.
The crash was caused by the attempt to call unmapped code: The subsystem
plugin libraries were getting unloaded before the execution of the subsystem
destructors.
Note:
In Android the issue is not seen since library unload is not implemented.

This patch ensures SystemClass destructor destroys the subsystems explicitely
unloading the plugins.

Change-Id: I19dad262b384bdbd63c7c319a41a5d547d0e75e9
Signed-off-by: Patrick Benavoli <patrick.benavoli@intel.com>