Fix test-platform set criterion by Lexical State Bug
BZ: 213233

The setCriterionState command of the Test Platform was
setting undesired value when asking for multiple states
to an InclusiveCriterion.
The method used in setCriterionStateByLexicalSpace was
to parse arguments before to get numerical value for
each of them with the getNumericalValue method from
ISelectionCriterionTypeInterface. Then, different
masks were aggregated to set the criterion's state.
Regarding the getNumericalValue method behaviour which
parses automatically value arguments, this method leads
to undesired bytes in the mask which leads to undesired
criterion's state.

The solution is to aggregate the setCriterionState
arguments when more than one arg are used.
As getNumericalValue use | to split his first parameter,
the code has been modified to use this behaviour.
getNumericalValue is now called only one
time and the returned mask is directly used to set
criterion's state.

Change-Id: Ie3a1d79a6f66928a2418f11233ff0d1ef5c3e900
Signed-off-by: Jules Clero <jules.clero@intel.com>