Move element tree initialisation to load function
BZ: 207091

The parameter platform manager used to initialize the element tree.
However, this is too early for plugins, since they need information from
the core which have not been initialized yet.

This patch delays the initialization of the element tree by adding it to
the load() function of the ParameterMgr.

Change-Id: I98e0e3a20ac0af2736eddc4c8349911eb1d1a691
Signed-off-by: Guillaume Denneulin <guillaume.denneulin@intel.com>
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>
Reviewed-on: https://android.intel.com/215426
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Reviewed-by: Benavoli, Patrick <patrick.benavoli@intel.com>
Reviewed-by: Rocard, KevinX <kevinx.rocard@intel.com>
Tested-by: Rocard, KevinX <kevinx.rocard@intel.com>