Separate Tests Action from Pfw Criteria Test Project
BZ: 209615

Previously, tests actions (such as media_playback) were
hardcoded in the PfwCriteriaTest project. The problem
was that the PFW can be used for much more than audio.
So, the PFWCriteriaTest project was only useful to test
audio use cases.

This patch aims to dissociate test cases generation
from test behaviour (such as media_playback).
To regroup all audio tests, a new configuration file
has been created : audio-test-type.json.

This file allows user to add new test behaviours.
So, each test vector is associated to a set of commands.
The associated script is launched when the execute
function of TestLauncher.py is called.
The audio-test-type.json files contains the type name,
its script, and a boolean to specify if we must wait the
entire script execution prior to continue.
This boolean is important when we want to launch long
commands, such as alsa commands that we want to run
during the entire test process.

Change-Id: Ic6cfcb145d4493de05f55f2d11be312621359663
Signed-off-by: Jules Clero <jules.clero@intel.com>
Reviewed-on: https://android.intel.com:443/240388