Add ability to generate an init state from criteria List
BZ: 209615

Previous version of PfwCriteriaTest project was asking for an
init file to be launched. It was not needed when init file was
corresponding to default values.

This patch allows PfwCriteriaTest to set default values when
no init file is given. The script just set the noValue to
all criteria.

The init file is now only useful to set a desired init
state to test the switch between two use cases.

Change-Id: Ib3b0928a7278f03189893ca9fb6e86835138891b
Signed-off-by: Jules Clero <jules.clero@intel.com>
Reviewed-on: https://android.intel.com:443/240389