Add Pfw coverage generation in Pfw Criteria Test Project
BZ: 209615

The parameter-framework comes with the coverage tool which
allows users to generate a complete report about domains
based on PFW log.

This patch adds the --coverage option to the Pfw Criteria
Test project which generates automatically a report based
on the script log.

The coverage script path, the domain file and the coverage
report file has been added to the conf.json file.

Change-Id: I21a4fa1d7ef96648c9e15462407eac1e65c7bc7c
Signed-off-by: Jules Clero <jules.clero@intel.com>
Reviewed-on: https://android.intel.com:443/240396