Enhance logging ability of external processes
BZ: 217046

When running external processes, PFW Criteria Test
project wait for the end of the process to display
stdout and sterr.
It can be a problem when using long script in
asynchronous mode.

This patch corrects this behaviour by using one thread
for each stream. So, each process is launched by a thread
which acts as manager of two StreamLoggerThread objects
which interact with launched subProcess.

The thread management has also be updated to let the
possibility to the user to wait the end of an
unfinished script by asking it to send SIGINT twice
if unfinished jobs are still running.

Change-Id: I5f93e4f00b1661c6fd56b9c1da1328deb628b863
Signed-off-by: Jules Clero <jules.clero@intel.com>
Reviewed-on: https://android.intel.com:443/240398