Update script usage declaration method
BZ: 219367

Pfw Criteria Test project allows users to launch
test scripts synchronously or asynchronously.

Previous versions were using a boolean present
in the 'ScriptsFile' to set the behaviour
of the script. It was complicated for new comers to
determine which value to use.

Now, each script should have one of the keyword:
asynchronous or synchronous to work.

Change-Id: Ia7791ed00764f1d589841381557b06699c7c3985
Signed-off-by: Jules Clero <jules.clero@intel.com>
Reviewed-on: https://android.intel.com:443/240399