Allow the plugin libraries to have empty path for location
If the xml defines the path of plugin library as empty, as a
slash is always added between the folder path and the library name,
the parameter framework will look for the library at the root of the
file system.

This patch prevents from adding the slash except if the path
is not empty.

Issue: GMINL-1748

Change-Id: I14ed5a5ddbef7e0305300479f4fd44bf611f3442
Signed-off-by: Louis Le Gall <louis.le.gall@intel.com>