fixed-point-tests: strip the '_host' suffix on executables
This only makes sense on the Android build system. However, these tests are
broken on Android because libremote-processor is compiled as
"libremote-processor_host.so" on Android whereas libparameter will only try to
load "libremote-processor.so" anyway.

For now, let's ignore the Android build system and only care about how we do it
with CMake: CMake builds libremote-processor.so, test-platform and
remote-process so let's remove the "_host" suffixes that were added for
compatibility with the output of the Android makefiles.

Signed-off-by: David Wagner <david.wagner@intel.com>